[18GB] Restrict buying president's certificate from the pool
In 18GB the president's certificate can be sold to the share pool, and
the corporation is placed in receivership. Once this has happened, a
player is not allowed to directly buy the president's certificate. They
must either buy two single shares (from either the IPO or the pool) or
buy a single share and then buy the president's certificate from the
pool and place the single share back in the pool.

This check was not in place and a player who did not own any shares in a
company was being given the option to buy the president's share from the
pool.

Fixes tobymao#11886.