[1822MX] remove `home_as_token: true` from Graph setup
`home_as_token: true` was added to 1822MX in #7554, citing ferry connections for minor acquisitions, but that functionality works fine in 1822PNW without `home_as_token: true`, and in 1822MX it is causing a bug where Majors with homes in Mexico City can reach hexes adjacent to Mexico City that they should not be able to reach.

Makes test added in previous comment pass. #TDD

Fixes #11148