[1822MX] remove `home_as_token: true` from Graph setup
`home_as_token: true` was added to 1822MX in #7554, citing ferry connections for
minor acquisitions, but that functionality works fine in 1822PNW without
`home_as_token: true`, and in 1822MX it is causing a bug where Majors with homes
in Mexico City can reach hexes adjacent to Mexico City that they should not be
able to reach.

(also clean up a comment from 1822 code with "hometokenplace" that was
copy-pasted 2x)

New fixture and spec file added in accordance with #11973

The update in the core `step/tracker.rb` makes a more readable `GameError` get
thrown if no hex neighbors are present, instead of a confusing `nil` error.

Fixes #11148