[1822MX/1822PNW] stop blocking on issue/redeem shares for P17 (#11980)
In 1822, P17 allows moving one of the cards in the auction stacks, and there is
a check to make sure this option is still available before the end of the owning
Major's turn.

In 1822MX, the ability of P17 is unrelated, and in fact there is not an
analogous private company. Fixture updates to remove unnecessary `pass` actions.

In 1822PNW, the analogous private (P21) can be owned minors, but they will still
skip the issue/redeem step since they don't have shares to issue/redeem and the
private can be used earlier in the turn anyway. Fixture updates to add `pass`
actions for Majors where the issue/redeem step now blocks.

1822CA and 1822Africa each already have a line setting `COMPANY_LUR` to `nil`.

Fixes #11048