[Core] Propogate `skip_track` when walking paths from a junction
The `skip_track` parameter was not being passed to subsequent
`Path.walk` calls after visiting a junction on a Lawson-style plain
track tile.

This was causing errors in the graph calculation for 1858 where a broad
gauge route could be calculated from a token to a broad or dual gauge
track tile, pass through the junction and then leave the tile onto a
narrow gauge path on another tile.

Fixes tobymao#12031.