[1861/1867] Override `game_end_set_final_turn!`
By overriding the `game_end_set_final_turn!` method the event triggering
the end of the game on the purchase/export of an 8 train can be removed.

This also includes a small tidy up to the 1867 code, using phase
statuses to check for a train being exported instead of hardcoding the
phase range this takes place in.