[core] replace `undefined method _ for nil` with new `ActionError`
raise `ActionError` when `REQUIRED_ARGS` are missing for action creation

While validating locally for #12086 and #12095, I happened to have local
games broken by #12073 (they are pinned in prod), and the error was very
unhelpful:

`#<NoMethodError: undefined method `id' for nil:NilClass>"`

Additionally, `broken_action.to_h` caused its own error in the validate
script, as the error was coming from the Action -> hash conversion.

With these changes, the error is much more helpful:

`#<Engine::ActionError: Cannot create Engine::Action::LayTile, h_to_args() returned nil :tile from action 745>`

----

In Ruby, [`Method#parameters`][0] can be used to inspect the arguments of a
and identify the required kwargs, so that could be used instead of defining
`REQUIRED_ARGS` for every action.

However, for this to work in Opal, the `Opal::Compiler` must have `arity_check:
true`.

This setting enforces that methods/lambdas/procs/functions are called with the
correct number of arguments in the JavaScript; after a few minutes of testing,
this appears to break most of our views, and fixing an error quickly leads to
the discovery of the next one.

JS event handlers, for example, are called with more args by the browser than we
care about or define. To make this work with Opal's arity check, dummy args need
to be added to every such function. It doesn't seem feasible to fix everything
in this way--which functions are affected and how many arguments are expected is
not clear until Opal throws the error, so fully implementing this check at this
point would be quite painful, and essentially impossible to roll out with
confidence that users wouldn't encounter the error.

[0]: https://docs.ruby-lang.org/en/master/Method.html#method-i-parameters