[1822PNW] Add impassable border between hexes O6 and O8
In the published edition of 1822PNW, Portland is represented by two grey
hexes, O8 and P9. Upgrades are done by placing a cylinder on the hexes.

Here Portland is a single hex, O8, that starts off as white and is
upgraded by changing the tile. There is a problem with this: there is
no track leaving to the west, but because O8 is not grey it is possible
to lay a tile in hex O6 that connects to O8. This shouldn't be allowed.

This commit adds an impassable border between O6 and O8 to prevent track
from being laid in this way.

Fixes tobymao#12123.

Seattle has also be changed from a grey hex to a regular hex, but this
does not present a problem as there are track connections from Seattle
to all neighouring land hexes.