[core] transform auto_actions attached to messages in `#to_h` and when scrubbing
* model change: when removing messages, if a message had auto actions attached,
  then the first auto action instead takes the place of the message

* scrube change: if a chat message has auto actions attached, replace the
  contents of the message with `"[redacted]"` so that the auto actions appear in
  the scrubbed DB as they do in the prod DB, but the user's chat content is
  still kept out of the scrubbed DB

Closes #12131

In current prod, accessing https://18xx.games/game/222776 appears broken if not
logged in as one of the players in that game. The API response keeps messages
hidden from you, but somehow a `pass` auto action got attached to a chat
message; with that `pass` action missing, the game cannot be processed
successfully by your browser.

This change doesn't prevent auto actions from attaching to chats again, but does
prevent games with such occurrences from appearing broken, and fixes consistency
when the game is opened in the scrubbed DB.