[core] make `slot` not required for `Action::PlaceToken`
Follow up to #12156; some actions had changes like this in that PR because of
fixtures; requiring `slot` broke no fixture tests but did cause some games to
fail validation once it was deployed.

1822CA's destination token step does not provide a `slot`, and it is not the
only step that processes `place_token` where the slot is not actually used
during the processing.

allows these games to be _un_pinned:

```
{"1822CA"=>
  [186175,
   189426,
   193258,
   195126,
   196533,
   197292,
   197804,
   199815,
   201899,
   203273,
   205006,
   205350,
   213385,
   216359,
   218093,
   219734,
   221155],
 "1822CA ERS"=>[201351, 204869, 207532, 215467, 220361, 220908]}
```