[core/dev] fix handling chats with auto actions and or undo targets
* model: set `message` to an empty string when the action must be included in
  the response

* view: if a chat message is an empty string, don't display it

* scrubbing: replace all chat messages with empty strings but keep the actions
  around

In #12148 I had found some `undo` actions that had an `action_id` pointing to an
action that seemed to not exist; turns out there were pointing at chat messages
that were filtered by `#to_h` in `models/game.rb`. This change solves that
problem and is a better solution than #12178 for auto_actions.