[18Ardennes] Allow any sales before starting a public company
Remove restrictions on selling shares before starting a public company.
Before this change the sale would only be allowed if the money from the
sale was used to launch the new company at a higher price. This could
lead to a situation where a new public company could not be started as
the player did not have any free certificate slots, but were not allowed
to sell any shares to free up a slot, as they already had enough money
to start the new company at the highest par price.

The implementation had been based on a misreading of section 5.3 of the
18Ardennes rules. This mentions using money to increase the par price,
but that is an example of why this might be done, not a requirement for
the sale to be allowed.

Fixes tobymao#12185.