[18Ardennes] Remove incorrect certificate limit check
Remove an unneeded and incorrect check against certificate limit when
buying shares after a public company is converted to a ten share
company.

There should not be a certificate limit check here as a player is
allowed to go above certificate limit during an operating round, and
only has to sell down if they are still above limit at the next stock
round.

The check is incorrect as it was actually checking the number of shares
in the converted company's treasury, not the number of certificates held
by the player.

Fixes tobymao#12225.