[1824] Don't skip action if there are company actions available
In a stock round, a player's turn was being skipped if there are no
actions returned by a call to `actions(entity)`, even if they own a
mountain railway company that could be exchanged for a share in a
regional railway. The actions for the companies are calculated in a
different method (`company_actions`) that is not checked when deciding
whether this step can be skipped.

This adds a check to `company_actions` in `blocking?`. It also adds some
more checks to `company_actions` to make sure that an action is only
returned if the mountain railway is owned by the active player, and they
have not already purchased a share in their current action.

Fixes tobymao#12258.