[1871] Remove a player from `pass_order` when they make a split action
Priority deal is based on passing order. If a player has passed in a
stock round they will be added to the `pass_order` array. If they then
make a split action they need to be removed from this array, so the
priority calculation is based on when the subsequently pass.

Fixes tobymao#12268.