[18Mag] Avoid infinite recursion in `red_to_red_route?`
There was a infinite loop occuring in games with the CIWR major company
when a train was converted into a plus train.

`red_to_red_route?` was checking `route.stops`. This is calculated by
calling `calculate_route` which, for a plus train, calls `revenue_for`.
And that method calls `red_to_red` route.

Avoid this problem by looking at `route.visited_stops` instead of
`route.stops`. The off-board hexes in 18Mag are all terminals, so we can
just check if the first and last stops are both red tiles.

This would fail if it is possible to have a route that visits an
off-board area but does not use it for its revenue. AFAIK this is not
possible in 18Mag.

Fixes tobymao#11443.