[View] Use relative height for timeline progress bar
The progress bar, shown in the timeline for some games, had a fixed
height of 55px. This is fine for most cases, but if the user has changed
their fonts to be larger than the default this could result in the text
not fitting in visible space.

Change this to use a relative height of 3.5em. This will grow or shrink
as the font size is changed.

Fixes tobymao#12303.