[1856/1836Jr56] Fix corporation capitalisation
In 1856 the changes to corporation capitalisation does not match up with
the phase changes: the change from escrow to incremental capitalisation
is when the first five train becomes available for purchase, and the
change from incremental to full capitalisation is when the first six
train becomes available.

This has been implemented with the 4' and 5' phases but these have been
eliminated. Instead track the capitalisation state in a
`@capitalization` variable, which is updated in events tied to the
purchase of the last 4-train and 5-train.