[Core/1856/1836Jr56] Hidden events that are not shown on game info page
An alternate implementation of hidden events. These are ones that are
not listed on the game info page.

Instead of a global `HIDDEN_EVENTS` array, this is done by adding a
`hidden` key to the event hash. If this is a truthy value then the event
will not be shown.