[1861/1867] Unlimited 8/2+2/5+5E trains
The last ranks of trains (8-trains, 2+2 trains and 5+5E trains) are
unlimited in both 1861 and 1867. These had been implemented by setting
up the train depot with a large number of trains, but tobymao#12236
added support for unlimited trains, so change these to be unlimited.

[Technically, unlimited 2+2 and 5+5E trains is a variant in 1861, but
it's the rules that everyone plays, and the base rule of two 5+5E trains
and two (or zero) 2+2 trains has not been implemented here].