[1870] Skip destination connect check if the game has ended
In the $400 finish variant, the game ends immediately once a corporation
reaches the last space on the stock market. This means that destination
checks should not take place after the operating turn in which this
space was reached.

This desination check was being done, as an auto action added to the
last action in the operating turn (the dividend step). This auto action
was being rejected with a `GameIsOver` error (and the game was then
getting stuck in an infinite loop, see tobymao#12315).

Fixes tobymao#12312, and adds a fixture and game spec tests for this
case.