[Core] Clone base variant of unlimited trains
If there is an unlimited rank of train with multiple variants, we must
always clone the base variant when creating a new train to add to the
depot. Without this change, purchasing a variant of the unlimited train
was resulting in a clone of the train variant being added to the depot,
with the base variant no longer available.